requirements.txt
APScheduler==3.10.0
asgiref==3.6.0
black==23.1.0
click==8.1.3
colorama==0.4.6
Django==4.1.7
djangorestframework==3.14.0
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
Pygments==2.14.0
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3-1.26.14
Werkzeug==2.2.3